+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 7.88
 -----
-
+- Add Recenlty Played sorting option for podcasts [#2978](https://github.com/Automattic/pocket-casts-ios/issues/2978)
 
 7.87
 -----